02_code/requirements.txt
qdrant-client>=1.12.0
openai>=1.68.0
python-dotenv>=1.0.1
unstructured[md]>=0.16.0